Move ProductForm to entities
