+++ src/frontend/src/entities/product/model/index.ts
@@ -2,3 +2,4 @@ export * from './constants';
 export * from './types';
 export * from './validation';
 export * from './productSlice';
+export * from './productSchema';
