+++ src/frontend/src/entities/product/ui/ProductForm.tsx
