+++ src/frontend/src/entities/product/ui/index.ts
@@ -1,1 +1,2 @@
+export * from './ProductForm';
 export * from './ProductInputForm';
