+++ src/frontend/src/features/manageNote/model/index.ts
@@ -4,4 +4,3 @@ export const { actions, selectors, reducer } = manageNoteSlice;
 
 export * from './types';
 export * from './noteSchema';
-export * from './productSchema';
