+++ src/frontend/src/shared/lib/index.ts
@@ -5,3 +5,4 @@ export * from './urlHelper';
 export * from './fileHelper';
 export * from './validation';
 export * from './inputMapping';
+export * from './quantitySchema';
